requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
drf-spectacular==0.27.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
intasend-python==1.0.1
requests==2.32.3
transformers==4.42.4
gunicorn==22.0.0
pytest==8.2.2
pytest-django==4.8.0
responses==0.25.3